requirements.txt
python-amazon-product-api
lxml
boto3
ebaysdk
